services/gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
redis==5.1.1
python-dotenv==1.0.1
africastalking==1.2.5
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
